Parte de Mari V2
Arreglado

+++ views/iconographies/view.ctp
@@ -8,12 +8,12 @@ echo $this->Html->script('zoomooz/jquery.zoomooz-zoomContainer');
 echo $this->Html->script('zoomooz/purecssmatrix');
 echo $this->Html->script('zoomooz/sylvester.src.stripped');
 echo $this->Html->css('website-assets/website');*/
-echo $this->Html->script('jquery.easing.1.3.js');
-echo $this->Html->script('turn');
+//echo $this->Html->script('jquery.easing.1.3.js');
+//echo $this->Html->script('turn');
 //echo $this->Html->script('wijmo/jquery.wijmo-open.all.2.2.1.min');
 //echo $this->Html->script('wijmo/jquery.wijmo-complete.all.2.2.1.min');
 //echo $this->Html->script('wijmo/jquery.wijmo.wijcarousel');
-echo $this->Html->script('bootstrap/bootstrap-tab');
+//echo $this->Html->script('bootstrap/bootstrap-tab');
 echo $this->Html->script('jquery.printPage.js');
 //echo $this->Html->script('pdfobject_source');
 
@@ -65,16 +65,34 @@ function marc21_decode($camp = null) {
 	}
 </style>
 <style type="text/css">
-#caja{
+	
+.marca-de-agua {
+    background: #000 url("/bvmjm/app/webroot/img/iconografia/logoprint.png") no-repeat bottom right;
+    width: 80px;
+    height: 341px;
+    margin: 0 auto;
+    display: block;
+    position: relative;
+    margin-top: 80px;
+}
+.marca-de-agua img{
+    padding:0;
+    width:100%;
+    height:100%; 
+    margin:0;
+    filter:alpha(opacity=80);
+    opacity:.80;
+}
+/*#caja{
    color: #fff;
    background-color: #fff;
    margin-top:-97px;
-   width: 480px;
+   width: 4px;
    margin-left: 10px;
    position: relative;
    opacity: 0.7;
-   height: '90px;
-}
+   height: 90px;
+}*/
 </style>
 <style>
 	.btn-primary {
@@ -103,23 +121,26 @@ function marc21_decode($camp = null) {
      	var pwin;
      	<?php
      	$title = marc21_decode($item['Item']['245']); ?>;
+     <?php	$author = marc21_decode($item['Item']['100']); ?>;
      	html =  '<html><head>';
      	html += '<style type="text/css">img { float: left; margin: 0; padding: 0;}</style>';
-     	html += '<style type="text/css">#bljaIMGte{float:left;position:relative}</style>';
+     	//html += '<style type="text/css">.marca-de-agua { background: #000 url("/bvmjm/app/webroot/img/iconografia/logoprint.png") no-repeat bottom right;  margin: 0 auto;display: block;position: relative;margin-top: 40px;}</style>';
 		html += '<style type="text/css">#bljaIMGte .bljaIMGtex{width:320px;position:absolute;top:540px;left:14px;}</style>';
      //	html += '<style type="text/css">#caja{color: #fff;background-color: #fff;width: 480px;margin-left: 10px;margin-top:550px;position: absolute;opacity: 0.7;height:90px}</style>';
      	//html += '<link rel="stylesheet" type="text/css" href="/webroot/css/abyayala/abyayala.css" />';
-     	html += '<link rel="stylesheet" type="text/css" href="/webroot/css/nuevo/bootstrap.min.css" /></head><body>';
+     	html += '<link rel="stylesheet" type="text/css" href="app/webroot/css/nuevo/bootstrap.min.css" /></head><body>';
      	html += '<div class="row">';
      	html += '<div class="container">';
-     	html += '<h1> <?php echo $title['a'] ?></h1>';
+     	html += '<h3> <?php echo $title['a'] ?></h3>';
+     	html += '<h4> <?php echo $author['a'] ?></h4>';
      	html += '<hr>';
-     	html += '<div id="bljaIMGte" >';
-        html += '<?php echo $this->Html->image('/webroot/covers/' . $item['Item']['cover_path'], array('width' => '500px', 'height'=>'600px', 'margin-left'=>'900px'))?>' ;
-        html += '<div id="caja" style= "color: #fff;background-color: #fff;;width: 502px;margin-left: -1px;margin-top:508px;position: absolute;opacity: 0.7;height:132px"><p style="text-align:center;color: #6C3F30;margin-top:-2px" ><b>Universidad Central de Venezuela </br>Biblioteca Virtual Musicológica "Juan Meserón"</b></p>';	
-        html += '<div id="img" style="margin-left: 130px;width:100px; height: 40px;margin-top:-20px;"> <?php echo $this->Html->image('/webroot/img/iconografia/logo5.png', array('display'=>'block','margin-right'=>'auto','margin-left'=>'auto','width'=> '175px', 'height'=>'46px', 'margin-top'=>'-20px'))?>';
+     	//html += '<div id="bljaIMGte" >';
+       	html += '<?php echo $this->Html->image('/app/webroot/covers/' . $item['Item']['cover_path'], array('max-width' => '90%','opacity'=>'.5','filter'=>'alpha(opacity=80)', 'height'=>'90%', 'margin-left'=>'1000px'))?>' ;
+       html += '<div id="caja" style= "color: #fff;background-color: transparent;width: 100px;margin-left: 10px;margin-top:515px;position: absolute;opacity: 0.6;height:70px"><p style="text-align:center;color: #6C3F30;margin-top:-510px" >';	
+        html += '<div id="img" > <?php echo $this->Html->image('/app/webroot/img/iconografia/logoprint1.png', array('display'=>'block','margin-right'=>'auto','margin-left'=>'auto', 'width'=> '60%', 'height'=>'50%'))?>';
+       // html += '<div class="marca-de-agua" style="width:80px !important; height:90px !important; margin-top: 500px">';
         html += '</div>';
-        html += '</div>';
+        //html += '</div>';
         html += '</div>';
         html += '</div>';
         html += '<br clear="both">';
@@ -135,12 +156,13 @@ function marc21_decode($camp = null) {
 </script>
 
 
-  
-<ul class="breadcrumb" style="margin: 0">
-  <li><a href="<?php echo $this->base; ?>">Inicio</a></li>
-  <li><a href="<?php echo $this->base; ?>/iconographies">Iconografía Musical Venezolana</a></li>
-  <li>
-  	<?php
+
+<?php if (($this->Session->check('Auth.User') && ($this->Session->read('Auth.User.group_id') == '2'))) { ?>
+<ul class="breadcrumb" style="margin: 0">	
+<li><font size="1.5" color="gray">Ir a</font></li>
+<li><a href="<?php echo $this->base; ?>/configurations">Inicio</a></li>
+ <li><a href="<?php echo $this->base; ?>/iconographies">Iconografía Musical en Venezuela</a></li>
+  <li>	<?php
 		if (!empty($item['Item']['245'])) {
 			$title = marc21_decode($item['Item']['245']);
 			if ($title) {
@@ -150,19 +172,55 @@ function marc21_decode($camp = null) {
 				if (isset($title['h'])) {echo ' ' . $title['h'];}
 			}
 		}
-	?>
-  	</li>
+	?></li>
 </ul>
+<?php } else if (($this->Session->check('Auth.User') && ($this->Session->read('Auth.User.group_id') == '1'))) { ?>
+<ul class="breadcrumb" style="margin: 0">	
+<li><font size="1.5" color="gray">Ir a</font></li>
+<li><a href="<?php echo $this->base; ?>/configurations">Inicio</a></li>
+ <li><a href="<?php echo $this->base; ?>/iconographies">Iconografía Musical en Venezuela</a></li>
+  <li>	<?php
+		if (!empty($item['Item']['245'])) {
+			$title = marc21_decode($item['Item']['245']);
+			if ($title) {
+				echo $title['a'];
+				if (isset($title['b'])) {echo ' ' . $title['b'];}
+				if (isset($title['c'])) {echo ' ' . $title['c'];}
+				if (isset($title['h'])) {echo ' ' . $title['h'];}
+			}
+		}
+	?></li>
+</ul>
+<?php } else { ?>
+<ul class="breadcrumb" style="margin: 0">	
+<li><font size="1.5" color="gray">Ir a</font></li>
+<li><a href="<?php echo $this->base; ?>/pages">Inicio</a></li>
+ <li><a href="<?php echo $this->base; ?>/iconographies">Iconografía Musical en Venezuela</a></li>
+  <li>	<?php
+		if (!empty($item['Item']['245'])) {
+			$title = marc21_decode($item['Item']['245']);
+			if ($title) {
+				echo $title['a'];
+				if (isset($title['b'])) {echo ' ' . $title['b'];}
+				if (isset($title['c'])) {echo ' ' . $title['c'];}
+				if (isset($title['h'])) {echo ' ' . $title['h'];}
+			}
+		}
+	?></li>
+</ul>
+<?php } ?>
+
+
 
 <div class='book view'>
 	<div class="col-md-2 column">
 	<br />
 		<div style="width: 100%; text-align: center;">
 			<?php
-				if (($item['Item']['cover_name']) && (file_exists($_SERVER['DOCUMENT_ROOT'] . "/".$this->base."/webroot/covers/" . $item['Item']['cover_path']))){
-					echo $this->Html->image("/webroot/covers/" . $item['Item']['cover_path'], array('width' => '90%','height' => '183px', 'title' => $item['Item']['cover_name']));
+				if (($item['Item']['cover_name']) && (file_exists($_SERVER['DOCUMENT_ROOT'] . "/".$this->base."/html/app/webroot/covers/" . $item['Item']['cover_path']))){
+					echo $this->Html->image("/app/webroot/covers/" . $item['Item']['cover_path'], array('width' => '90%','height' => '183px', 'title' => $item['Item']['cover_name']));
 				} else {
-					echo $this->Html->image("/webroot/img/sin_portada.jpg", array('width' => '90%'));
+					echo $this->Html->image("/app/webroot/img/sin_portada.jpg", array('width' => '90%'));
 				}
 			?>
 		</div>
@@ -187,13 +245,13 @@ function marc21_decode($camp = null) {
 				?>
 				</dd>
 				<dt><?php __('Publicación'); ?>:</dt>
-				<dd>
+				<dd style="margin-left: 180px; text-align: justify; width:65%">
 					<?php
 						if (!empty($item['Item']['260'])) {
 							$year_century = marc21_decode($item['Item']['260']);
 							echo $year_century['a'] . '.';
-							if (isset($year_century['b'])) {echo '<br />' . $year_century['b']. '.';}
-							if (isset($year_century['c'])) {echo '<br />' . $year_century['c']. '.';}
+							if (isset($year_century['b'])) {echo " : " . $year_century['b']. ', ';}
+							if (isset($year_century['c'])) {echo $year_century['c']. '. ';}
 						}
 					?>
 				</dd>
@@ -236,12 +294,14 @@ function marc21_decode($camp = null) {
 					<?php
 						if (!empty($item['Item']['650'])) {
 							$mattername = marc21_decode($item['Item']['650']);
-							echo $mattername['a'];
+							echo $mattername['a']. '.';
 						}
+							if (isset($mattername['x'])) {echo ' ' . $mattername['x']. '.';}
 					?>
 				</dd>
 				<?php } ?>
 				
+							
 				<?php if (!empty($item['Item']['773'])) { ?>
 				<dt><?php __('Fuente'); ?>:</dt>
 				<dd>
@@ -249,6 +309,17 @@ function marc21_decode($camp = null) {
 						if (!empty($item['Item']['773'])) {
 							$source = marc21_decode($item['Item']['773']);
 							echo $source['t'];
+						}
+					?>
+				</dd>
+				<?php } ?>
+				<?php if (!empty($item['Item']['500'])) { ?>
+				<dt><?php __('Descripción'); ?>:</dt>
+				<dd style="margin-left: 180px; text-align: justify; width:65%">
+					<?php
+						if (!empty($item['Item']['500'])) {
+							$description= marc21_decode($item['Item']['500']);
+							echo $description['a'];
 						}
 					?>
 				</dd>
@@ -276,49 +347,34 @@ function marc21_decode($camp = null) {
 				if (($this->Session->check('Auth.User') && ($this->Session->read('Auth.User.group_id') != '3'))) {
 					echo $this->Html->link('Agregar Obra', array('action' => '/add'), array('class' => 'btn-primary', 'title' => 'Agregar Obra'));
 					echo $this->Html->link('Modificar Obra', array('action' => '/edit/'.$item['Item']['id']), array('class' => 'btn-primary', 'title' => 'Modificar Obra'));
-				}
-			?>
-			<?php
+					echo $this->Html->link(__('Eliminar Obra', true), array('action' => 'delete', $item['Item']['id']), array('class' => 'btn-primary', 'title' => 'Eliminar Obra'), sprintf(__('Are you sure you want to delete # %s?', true), $title['a']));
+				}else
+				if (($this->Session->check('Auth.User') && ($this->Session->read('Auth.User.group_id') == '3'))) {
 				echo $this->Form->hidden('user_id', array('type' => 'text', 'value' => $this->Session->read('Auth.User.id')));
 				echo $this->Form->hidden('item_id', array('type' => 'text', 'value' => $item['Item']['id']));
 				echo $this->Html->link('Agregar a Mi Biblioteca', array('action' => '#'), array('id' => 'biblioteca', 'class' => 'btn-primary', 'title' => 'Agregar a Mi Biblioteca', 'onclick' => 'return false;'));
+				}else
+				if (($this->Session->check('Auth.User') && ($this->Session->read('Auth.User.group_id') == '1'))) {
+					echo $this->Html->link(__('Eliminar Obra', true), array('action' => 'delete', $item['Item']['id']), array('class' => 'btn-primary', 'title' => 'Eliminar Obra'), sprintf(__('Are you sure you want to delete # %s?', true), $title['a']));
+				}
+			?>
+			<?php
+				
 				echo $this->Html->link('Ver Formato MARC21', array('action' => 'marc21/'.$item['Item']['id']), array('class' => 'btn-primary', 'title' => 'Formato MARC21'));
 			?>
 			<?php if (!empty($item['Item']['cover_path'])) { ?>
-				<a href="http://<?php echo $_SERVER['HTTP_HOST'] . $this->base . '/webroot/covers/' . $item['Item']['cover_path']; ?>" class="btn-primary" target="_blank" title="Descargue el documento en su computadora.">Descargar Imagen</a>
+				<a href="http://<?php echo $_SERVER['HTTP_HOST'] . $this->base . '/app/webroot/covers/' . $item['Item']['cover_path']; ?>" class="btn-primary" target="_blank" title="Descargue el documento en su computadora.">Descargar Imagen</a>
 			<?php } ?>
 			
 			<?php if (!empty($item['Item']['cover_path'])) { ?>
 				<a href="#" onclick="return imprimir();" class="btn-primary" target="_blank" title="Imprimir Imagen.">Imprimir Imagen</a>
-			<?php } ?>	
+			
+			<?php } ?>		
+			
 		</form>
 	</div>
 	
-	<div style="clear: both;"><br /></div>
-	
-	<?php if ($item['Item']['item_content_type'] == "application/pdf") { ?>
-		<?php if ($item['Item']['item_file_path']) { ?>
-			<!-- <iframe src="<?php echo 'http://' . $_SERVER['HTTP_HOST'] . $this->base . '/webroot/files/' . $item['Item']['item_file_path']; ?>" width="99%" height="600px"></iframe> -->
-			<!-- <iframe src="http://docs.google.com/viewer?url=<?php echo 'http://' . $_SERVER['HTTP_HOST'] . $this->base . '/webroot/files/' . $item['Item']['item_file_path']; ?>" width="99%" height="600px"></iframe> -->
-			<!-- <object width="99%" height="600" type="application/pdf" data="<?php echo 'http://' . $_SERVER['HTTP_HOST'] . $this->base . '/webroot/files/' . $item['Item']['item_file_path']; ?>">
-			<param name="src" value="<?php echo 'http://' . $_SERVER['HTTP_HOST'] . $this->base . '/webroot/files/' . $item['Item']['item_file_path']; ?>" />
-			<p>N o PDF available</p>
-			</object> -->
-			
-			<object data="<?php echo 'http://' . $_SERVER['HTTP_HOST'] . $this->base . '/webroot/files/' . $item['Item']['item_file_path']; ?>" type="application/pdf" width="100%" height="600px">
-			<br /><br />
-			<div style="text-align: center;">
-				Lamentablemente este navegador no posee un plugin para visualizar PDF's.
-			
-</div>
-			<br />
-				Instale un plugin para visualizar el PDF o descargue el archivo <a href="http://<?php echo $_SERVER['HTTP_HOST'] . $this->base . '/webroot/files/' . $item['Item']['item_file_path']; ?>" target="_blank" title="Descargue el documento en su computadora.">aquí</a>. 
-			<br /><br /><br /><br />
-			</div>
-			</object>
-			
-		<?php } ?>
-	<?php } else {echo "<div style='text-align: center'>El archivo no tiene formato pdf.</div><br />";} ?>
+
 </div>
 
 <script type="text/javascript">
